installed react-draft-wysiwyg (text editor), customized it for new comment section, added NewComment component and AddCommentBtn component, moved NewThread to forums folder
