+++ package.json
@@ -7,9 +7,11 @@
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
     "bootstrap": "^4.6.0",
+    "draft-js": "^0.11.7",
     "react": "^17.0.2",
     "react-bootstrap": "^1.6.1",
     "react-dom": "^17.0.2",
+    "react-draft-wysiwyg": "^1.14.7",
     "react-dropzone": "^11.3.2",
     "react-icons": "^4.2.0",
     "react-multiselect-checkboxes": "^0.1.1",
