+++ src/components/Forum/AddCommentBtn.js
@@ -1,0 +1,9 @@
+import React from 'react'
+import { MdSend } from 'react-icons/md'
+export default function AddCommentBtn() {
+    return (
+        <div>
+            <MdSend />
+        </div>
+    )
+}
